# temp_thkiengiangshoes/thkiengiangshoes-main/docs/openapi.yaml
openapi: 3.0.3
info:
  title: TBS II Production & Maintenance Management System API
  description: OpenAPI Specification for TBS II Multi-platform Factory Operations, SLA Calculation Engine, and BI Dashboard.
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: Local Development API Server
paths:
...
  /incidents:
    get:
      summary: List Incidents
      tags: [Incidents]
      responses:
        '200':
          description: List of reported incidents with SLA metrics and downtime
    post:
      summary: Report New Machine Incident (Worker)
      tags: [Incidents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [machine_id, description]
              properties:
                machine_id: { type: integer, example: 3 }
                priority: { type: string, example: "CRITICAL" }
                description: { type: string, example: "Máy cắt dừng đột ngột" }
      responses:
        '200':
          description: Incident created, machine status set to DOWN, real-time alert broadcast
  /incidents/{id}/accept:
    post:
      summary: Accept Maintenance Task (Starts SLA Stopwatch)
      tags: [Incidents]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Task accepted, accepted_at timestamp recorded, response time calculated
  /incidents/{id}/resolve:
    post:
      summary: Confirm Repair Complete (Maintenance)
      tags: [Incidents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [incident_id, root_cause, resolution_notes]
              properties:
                incident_id: { type: integer }
                root_cause: { type: string }
                resolution_notes: { type: string }
                spare_parts_used: { type: string }
      responses:
        '200':
          description: Incident resolved, downtime calculated, machine restored to OPERATING
...